+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 * Adding support for multiple emitters per collector [#28](https://github.com/getoutreach/conrad/pull/28)
+
+## Version 2.4.0
 * Adding the ability to emit events in a background thread [#28](https://github.com/getoutreach/conrad/pull/28)
 * Adding a Kinesis Emitter [#30](https://github.com/getoutreach/conrad/pull/30)